Enhance user management: implement profile check middleware, update user policy for show action, and refine user routes

+++ my-project/routes/web.php
@@ -32,11 +32,11 @@ Route::get('/dashboard', [App\Http\Controllers\HomeController::class, 'index'])-
 
 // Admin Users route
 Route::prefix('/admin/users/')->name('user.')->group(function () {
-    Route::get('index', [UserController::class, 'index'])->name('index');
-    Route::get('{user}/show', [UserController::class, 'show'])->name('show');
-    Route::get('{user}/edit', [UserController::class, 'edit'])->name('edit');
-    Route::put('{user}/update', [UserController::class, 'update'])->name('update');
-    Route::delete('{user}/delete', [UserController::class, 'destroy'])->name('destroy');
-    Route::post('{user}/restore', [UserController::class, 'restore'])->name('restore');
-    Route::delete('{user}/force-delete', [UserController::class, 'forceDelete'])->name('forceDelete');
+    Route::get('index', [UserController::class, 'index'])->name('index')->middleware('superadmin');
+    Route::get('{user}/show', [UserController::class, 'show'])->name('show')->middleware('check.profile');
+    Route::get('{user}/edit', [UserController::class, 'edit'])->name('edit')->middleware('check.profile');
+    Route::put('{user}/update', [UserController::class, 'update'])->name('update')->middleware('check.profile');
+    Route::delete('{user}/delete', [UserController::class, 'destroy'])->name('destroy')->middleware('superadmin');
+    Route::post('{user}/restore', [UserController::class, 'restore'])->name('restore')->middleware('superadmin');
+    Route::delete('{user}/force-delete', [UserController::class, 'forceDelete'])->name('forceDelete')->middleware('superadmin');
 });
